requirements.txt
astroid==2.9.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.5.1
pylint==2.12.2
PyYAML==6.0
toml==0.10.2
wrapt==1.13.3
